Simplify: rename regex_question module now we don't have that named struct
